// domain/content/types.ts
export interface Step {
  id: number
  title: string
  description: string
}

export interface Advantage {
  id: number
  title: string
  description: string
}

export interface Feature {
  id: number
  title: string
  description: string
  icon: string
}

export interface UseCase {
  id: number
  title: string
  description: string
  icon: string
}

export interface FAQ {
  id: number
  question: string
  answer: string
}